mplayer: switch back to video PTS for reporting playback time
The main problem with video PTS was that it wasn't very useful when
playing audio files with cover art. Using the audio time instead was an
obvious solution. Unfortunately, this leads to "inexact" reporting of
the playback time in paused mode, and audio is always ahead by small,
essentially random amounts of time ahead. This is possibly because the
times reported by AOs are not entirely accurate when paused (see commit
9b3bf76).

Switch back to video PTS, and use a simpler way to deal with the cover
art case: if the video has ended, use the audio PTS.

Also see commit f9a259e (and the commits referenced from there).