configure: add -mconsole on MinGW
At least libsdl adds -mwindows to the cflags, which marks the .exe
binary as GUI application. This means the program detaches from the
console when started in cmd.exe, instead of showing the playback
status, receiving console input, and so on.

Append -mconsole to the linker command line to disable -mwindows.