add a primitive example
mpv somefile.mkv -lua test.lua

Replicates the OSD bar (sort of), and shows a '+' at the current mouse
position. (It's a centered ASCII '+' so the position doesn't have to be
completely exact.) The OSD bar is shown on mouse move. Clicking inside
the bar does a seek.

This is fully independent from whatever mpv does, for the better or the
worse.

This is just a demo, nothing actually useful.